requirements.txt
Django
gunicorn
django-jazzmin
django-ckeditor
python-dotenv
Pillow
